move plotting funcs to rasterio.plot
